Moving ExampleCsvPojo from src to test
